Split mullvad_daemon into lib.rs and main.rs
